+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Some english translations.
 - New `stage` argument in `odm_ui.UIMixin.setup_m_form()` hook.
 - New `args` argument in `core.odm.Model.t()`
+- New events in ODM model: `odm.entity.pre_save`, `odm.entity.save` 
+- New events in Content ODM model: `content.entity.pre_save`, `content.entity.save`
+- New properties in Content ODM model: `title`, `description`, `body`, `url`, `tags` 
+- New properties in Taxonomy ODM model: `title`, `alias`, `language`, `weight`, `order` 
 
 ### Fixed
 - Empty configuration files read error.